add zprofile check

+++ scripts/bootstrap_mac.sh
@@ -21,7 +21,11 @@ function main {
 
   echo "${BLUE}Installing zsh dotiles${NORMAL}"
   grep --fixed-strings "dotfiles/config/shell/init.sh" "$HOME"/.zshrc || mv "$HOME"/.zshrc "$HOME"/.zshrc.bak && cp "$HOME"/dotfiles/config/shell/mac/zshrc.sh "$HOME"/.zshrc
-  mv "$HOME"/.zprofile "$HOME"/.zprofile.bak && cp "$HOME"/dotfiles/config/shell/mac/zprofile.sh "$HOME"/.zprofile
+  if [ -e "$HOME"/.zprofile ]; then
+    mv "$HOME"/.zprofile "$HOME"/.zprofile.bak && cp "$HOME"/dotfiles/config/shell/mac/zprofile.sh "$HOME"/.zprofile
+  else
+    cp "$HOME"/dotfiles/config/shell/mac/zprofile.sh "$HOME"/.zprofile
+  fi
 
   echo "${BLUE}Installing zsh plugins${NORMAL}"
   ZSH_CUSTOM=${ZSH_CUSTOM:-~/.oh-my-zsh/custom}
@@ -49,7 +53,7 @@ function main {
     echo "unknown: $SYSTEM_ARCH"
     ;;
   esac
-  brew bundle --file="$HOME"/dotfiles/assets/brew/brew_dev.txt
+  brew bundle --file="$HOME"/dotfiles/assets/brew/Brewfile_dev
 
   echo "${BLUE}Installing mackup${NORMAL}"
   ln -sf "$HOME"/dotfiles/config/mackup/.mackup.cfg "$HOME"/.mackup.cfg
@@ -91,9 +95,9 @@ function main {
   echo "${BLUE}Installing SpaceVim${NORMAL}"
   curl -sLf https://spacevim.org/install.sh | bash
 
-  # echo "${BLUE}Installing doom-emacs${NORMAL}"
-  # git clone --depth 1 https://github.com/hlissner/doom-emacs "$HOME"/.emacs.d
-  # "$HOME"/.emacs.d/bin/doom install
+  echo "${BLUE}Installing Doom Emacs${NORMAL}"
+  git clone --depth 1 https://github.com/doomemacs/doomemacs ~/.emacs.d
+  ~/.emacs.d/bin/doom install
 
   asdf reshim
 }
